Created/Started: The online single-player application (the backbone of the game) has been launched. Attention: tcp-net library not less than 1.1 for server abd for client

+++ src/main/java/telran/game/bulls_cows/exceprions/GameNotFoundException.java
@@ -3,7 +3,7 @@ package telran.game.bulls_cows.exceprions;
 import java.util.NoSuchElementException;
 
 public class GameNotFoundException extends NoSuchElementException {
-    public GameNotFoundException(int gameId) {
+    public GameNotFoundException(Long gameId) {
         super("The game [id=" + gameId + "] is not found.");
     }
 }
